separated user data type into own file
