+++ fireflight/src/context/GlobalContext.jsx
@@ -1,8 +1,8 @@
-import React, { useReducer } from "react";
+import React, { useReducer } from 'react'
 // import FireContext from "./contextProvider";
-import { GlobalContext, defaultValues } from "./contextProvider";
+import { GlobalContext, defaultValues } from './contextProvider'
 
-import { SET_LOCATION, SET_NAME } from "./types";
+import { SET_LOCATION, SET_NAME } from './globalDataTypes'
 
 // REDUCER EXPLANATION:
 // We use a reducer for the same reason we would use it in redux.
@@ -15,18 +15,18 @@ const globalReducer = (state, action) => {
       return {
         ...state,
         location: action.payload
-      };
+      }
     case SET_NAME:
       return {
         ...state,
         name: action.payload
-      };
+      }
     default:
       return {
         ...state
-      };
+      }
   }
-};
+}
 
 // CREATE CONTEXT EXPLANATION:
 /* We initialize FireContext as an empty createContext object. We don't want to 
@@ -44,7 +44,7 @@ function GlobalProvider(props) {
   // This puts all of our variables into the state object.
   // This allows us to send the entire state object into the reducer to be properly updated.
 
-  const [state, dispatch] = useReducer(globalReducer, defaultValues);
+  const [state, dispatch] = useReducer(globalReducer, defaultValues)
 
   // SET HOOKS EXPLANATION:
   // The concept of the set functions is exactly the same as in a regular hook. We use the set function to set the data inside the state. These functions (think redux actions) use dispatch to pass the newly set data into the reducer. State is then updated properly.
@@ -53,11 +53,11 @@ function GlobalProvider(props) {
     dispatch({
       type: SET_NAME,
       payload: newUser
-    });
-  };
-  const setToken = newToken => {};
+    })
+  }
+  const setToken = newToken => {}
 
-  const setLocation = newLocation => {};
+  const setLocation = newLocation => {}
 
   //structure
   /**
@@ -110,7 +110,7 @@ function GlobalProvider(props) {
     >
       {props.children}
     </GlobalContext.Provider>
-  );
+  )
 }
 
-export default GlobalProvider;
+export default GlobalProvider
