+++ fireflight/src/context/globalDataTypes.js
