+++ fireflight/src/styles/App.scss
@@ -15,7 +15,7 @@
 @import 'public-map';
 @import 'map-legend';
 @import 'model';
-@import 'model-splits';
+// @import 'model-splits';
 @import 'dashboard';
 @import 'address-form';
